fixed test cases

+++ core/src/test/scala/com/jabong/dap/campaign/traceablility/PastCampaignCheckTest.scala
@@ -45,7 +45,7 @@ class PastCampaignCheckTest extends FlatSpec with SharedSparkContext {
   "Past campaign Data with 47 mail type" should "return no customer whom we have sent the campaign" in {
     val format = new SimpleDateFormat(TimeConstants.DATE_TIME_FORMAT)
     val date = format.parse("2015-07-09 00:00:08.0")
-    val ndays = TimeUtils.daysFromToday(date).toInt
+    val ndays = TimeUtils.daysFromToday(date)
     val mailTypeCustomers = pastCampaignCheck.getCampaignCustomers(pastCampaignData, 47, ndays)
     assert(mailTypeCustomers.count == 0)
   }
@@ -53,7 +53,7 @@ class PastCampaignCheckTest extends FlatSpec with SharedSparkContext {
   "Past campaign Data with 46 mail type" should "return one customer whom we have sent the campaign" in {
     val format = new SimpleDateFormat(TimeConstants.DATE_TIME_FORMAT)
     val date = format.parse("2015-07-09 00:00:08.0")
-    val ndays = TimeUtils.daysFromToday(date).toInt
+    val ndays = TimeUtils.daysFromToday(date)
     val mailTypeCustomers = pastCampaignCheck.getCampaignCustomers(pastCampaignData, 46, ndays)
     assert(mailTypeCustomers.count == 1)
   }
@@ -81,7 +81,7 @@ class PastCampaignCheckTest extends FlatSpec with SharedSparkContext {
   "Past campaign Data with 46 mail type to past campaign check" should "return one customer whom we have not sent the campaign" in {
     val format = new SimpleDateFormat(TimeConstants.DATE_TIME_FORMAT)
     val date = format.parse("2015-07-09 00:00:08.0")
-    val ndays = TimeUtils.daysFromToday(date).toInt
+    val ndays = TimeUtils.daysFromToday(date)
     val campaignNotSendCustomers = pastCampaignCheck.campaignCheck(pastCampaignData, customerSelected, 46, ndays)
     assert(campaignNotSendCustomers.count == 1)
   }
@@ -89,7 +89,7 @@ class PastCampaignCheckTest extends FlatSpec with SharedSparkContext {
   "Past campaign Data with 47 mail type to past campaign check" should "return two customer whom we have  not sent the campaign" in {
     val format = new SimpleDateFormat(TimeConstants.DATE_TIME_FORMAT)
     val date = format.parse("2015-07-09 00:00:08.0")
-    val ndays = TimeUtils.daysFromToday(date).toInt
+    val ndays = TimeUtils.daysFromToday(date)
     val campaignNotSendCustomers = pastCampaignCheck.campaignCheck(pastCampaignData, customerSelected, 47, ndays)
     assert(campaignNotSendCustomers.count == 2)
   }
